refactor API reference templates, split up long template files
