+++ package.json
@@ -73,6 +73,7 @@
     "eslint-plugin-prettier": "^3.0.0",
     "markdownlint-cli": "^0.13.0",
     "npm-run-all": "^4.1.5",
+    "ora": "^3.0.0",
     "prettier": "^1.15.3",
     "prettier-stylelint": "^0.4.2",
     "stylelint": "^9.9.0",
