+++ src/templates/Doc.module.scss
@@ -48,3 +48,9 @@
     max-width: 73%;
     margin: auto;
 }
+
+.version {
+    font-size: $font-size-base;
+    font-family: $font-family-monospace;
+    color: $brand-grey-light;
+}
