+++ src/templates/Swagger/Paths.module.scss
@@ -1,52 +1,5 @@
 @import 'variables';
 
-.version {
-    font-size: $font-size-base;
-    font-family: $font-family-monospace;
-    color: $brand-grey-light;
-}
-
-.swaggerMeta {
-    margin-top: -($spacer);
-    padding: 0;
-    font-size: $font-size-small;
-
-    li {
-        display: inline-block;
-        margin-left: $spacer;
-
-        &:first-child {
-            margin-left: 0;
-        }
-
-        &:before {
-            display: none;
-        }
-    }
-}
-
-.basePath {
-    margin-top: $spacer;
-
-    h2 {
-        font-size: $font-size-h3;
-        border-bottom: 1px solid $brand-grey-lighter;
-        padding-bottom: $spacer / 2;
-        margin-top: $spacer * 2;
-        margin-bottom: $spacer;
-    }
-
-    span {
-        color: $brand-grey-light;
-    }
-
-    code {
-        // stylelint-disable-next-line
-        font-size: $font-size-large !important;
-        font-weight: $font-weight-bold;
-    }
-}
-
 .pathName {
     font-size: $font-size-h3;
     border-bottom: 1px solid $brand-grey-lighter;
