+++ src/templates/Swagger/index.module.scss
@@ -1,0 +1,42 @@
+@import 'variables';
+
+.swaggerMeta {
+    margin-top: -($spacer);
+    padding: 0;
+    font-size: $font-size-small;
+
+    li {
+        display: inline-block;
+        margin-left: $spacer;
+
+        &:first-child {
+            margin-left: 0;
+        }
+
+        &:before {
+            display: none;
+        }
+    }
+}
+
+.basePath {
+    margin-top: $spacer;
+
+    h2 {
+        font-size: $font-size-h3;
+        border-bottom: 1px solid $brand-grey-lighter;
+        padding-bottom: $spacer / 2;
+        margin-top: $spacer * 2;
+        margin-bottom: $spacer;
+    }
+
+    span {
+        color: $brand-grey-light;
+    }
+
+    code {
+        // stylelint-disable-next-line
+        font-size: $font-size-large !important;
+        font-weight: $font-weight-bold;
+    }
+}
