+++ src/templates/Swagger/utils.js
@@ -1,0 +1,13 @@
+export const cleanPathKey = key => {
+    let keyCleaned = key
+
+    if (key.includes('aquarius')) {
+        keyCleaned = key.replace(/\/api\/v1\/aquarius/gi, '')
+    }
+
+    if (key.includes('brizo')) {
+        keyCleaned = key.replace(/\/api\/v1\/brizo/gi, '')
+    }
+
+    return keyCleaned
+}
